Release the 3.0.3 version
