+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [3.0.3] (2021-04-26)
+
+- DragonFly BSD support has been added. (#256)
+
+- Compilation for the FreeBSD operating system has been fixed. (#258)
+
 ## [3.0.2] (2021-04-05)
 
 - NixOS support has been added. (#252)
@@ -201,7 +207,8 @@ All notable changes to this project will be documented in this file.
 
 The first release containing only minor infrastructural changes and based on [os_type](https://github.com/schultyy/os_type).
 
-[Unreleased]: https://github.com/stanislav-tkach/os_info/compare/v3.0.2...HEAD
+[Unreleased]: https://github.com/stanislav-tkach/os_info/compare/v3.0.3...HEAD
+[3.0.3]: https://github.com/stanislav-tkach/os_info/compare/v3.0.2...v3.0.3
 [3.0.2]: https://github.com/stanislav-tkach/os_info/compare/v3.0.1...v3.0.2
 [3.0.1]: https://github.com/stanislav-tkach/os_info/compare/v3.0...v3.0.1
 [3.0.0]: https://github.com/stanislav-tkach/os_info/compare/v2.0.8...v3.0
