+++ cli/Cargo.toml
@@ -17,7 +17,7 @@ name = "os_info"
 path = "src/main.rs"
 
 [dependencies]
-os_info = { version = "3.0.2", default-features = false, path = "../os_info" }
+os_info = { version = "3.0.3", default-features = false, path = "../os_info" }
 log = "0.4.5"
 env_logger = "0.8.1"
 structopt = "0.3"
