+++ os_info/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "os_info"
-version = "3.0.2"
+version = "3.0.3"
 authors = ["Jan Schulte <hello@unexpected-co.de>", "Stanislav Tkach <stanislav.tkach@gmail.com>"]
 description = "Detect the operating system type and version."
 documentation = "https://docs.rs/os_info"
